+++ atoum/admin/__init__.py
@@ -1,17 +1,21 @@
-from .assortment import AssortmentAdmin
+from .assortment import AssortmentAdmin, AssortmentResource
 from .brand import BrandAdmin
-from .category import CategoryAdmin
-from .consumable import ConsumableAdmin
-from .product import ProductAdmin, ProductAdminInline
+from .category import CategoryAdmin, CategoryResource
+from .consumable import ConsumableAdmin, ConsumableResource
+from .product import ProductAdmin, ProductAdminInline, ProductResource
 from .shopping import ShoppingAdmin
 
 
 __all__ = [
     "AssortmentAdmin",
+    "AssortmentResource",
     "BrandAdmin",
     "CategoryAdmin",
+    "CategoryResource",
+    "ConsumableResource",
     "ConsumableAdmin",
     "ProductAdmin",
+    "ProductResource",
     "ProductAdminInline",
     "ShoppingAdmin",
 ]