+++ atoum/admin/consumable.py
@@ -1,12 +1,28 @@
 from django.contrib import admin
 from django.utils.translation import gettext_lazy as _
 
+from import_export import resources
+from import_export.admin import ImportExportModelAdmin
+
 from ..models import Consumable
 from .assortment import AssortmentInline
 
 
+class ConsumableResource(resources.ModelResource):
+    """
+    Data resource configuration for 'django-import-export'.
+    """
+    class Meta:
+        model = "atoum.Consumable"
+        fields = ("title", "slug")
+        import_id_fields = ("slug",)
+
+
 @admin.register(Consumable)
-class ConsumableAdmin(admin.ModelAdmin):
+class ConsumableAdmin(ImportExportModelAdmin):
+    """
+    Model admin controller.
+    """
     readonly_fields = ["created", "modified"]
     prepopulated_fields = {
         "slug": ("title",),
@@ -23,6 +39,7 @@ class ConsumableAdmin(admin.ModelAdmin):
     inlines = [
         AssortmentInline,
     ]
+    resource_classes = [ConsumableResource]
 
     def count_assortments(self, obj):
         """