+++ atoum/admin/shopping.py
@@ -5,6 +5,10 @@ from ..models import Shopping, ShoppingItem
 
 
 class ShoppingItemInlineAdmin(admin.StackedInline):
+    """
+    Inline model admin controller for 'ShoppingItem' to be used in other model admin
+    controllers.
+    """
     form = ShoppingItemInlineForm
     model = ShoppingItem
     exclude = ["created", "modified"]
@@ -14,6 +18,9 @@ class ShoppingItemInlineAdmin(admin.StackedInline):
 
 @admin.register(Shopping)
 class ShoppingAdmin(admin.ModelAdmin):
+    """
+    Model admin controller for 'Shopping'.
+    """
     inlines = [
         ShoppingItemInlineAdmin,
     ]